+++ lib/src/either.dart
@@ -1,3 +1,5 @@
+import 'dart:async';
+
 typedef Lazy<T> = T Function();
 
 /// Represents a value of one of two possible types.
@@ -35,14 +37,14 @@ abstract class Either<L, R> {
 
   /// Transform value of [Right] when transformation may be finished with an error
   Future<Either<L, TR>> thenAsync<TR>(
-      Future<Either<L, TR>> Function(R right) fnR);
+      FutureOr<Either<L, TR>> Function(R right) fnR);
 
   /// Transform value of [Left] when transformation may be finished with an [Right]
   Either<TL, R> thenLeft<TL>(Either<TL, R> Function(L left) fnL);
 
   /// Transform value of [Left] when transformation may be finished with an [Right]
   Future<Either<TL, R>> thenLeftAsync<TL>(
-      Future<Either<TL, R>> Function(L left) fnL);
+      FutureOr<Either<TL, R>> Function(L left) fnL);
 
   /// Transform value of [Right]
   Either<L, TR> map<TR>(TR Function(R right) fnR);
@@ -51,10 +53,10 @@ abstract class Either<L, R> {
   Either<TL, R> mapLeft<TL>(TL Function(L left) fnL);
 
   /// Transform value of [Right]
-  Future<Either<L, TR>> mapAsync<TR>(Future<TR> Function(R right) fnR);
+  Future<Either<L, TR>> mapAsync<TR>(FutureOr<TR> Function(R right) fnR);
 
   /// Transform value of [Left]
-  Future<Either<TL, R>> mapLeftAsync<TL>(Future<TL> Function(L left) fnL);
+  Future<Either<TL, R>> mapLeftAsync<TL>(FutureOr<TL> Function(L left) fnL);
 
   /// Fold [Left] and [Right] into the value of one type
   T fold<T>(T Function(L left) fnL, T Function(R right) fnR);
@@ -101,7 +103,7 @@ class Left<L, R> extends Either<L, R> {
 
   @override
   Future<Either<L, TR>> thenAsync<TR>(
-      Future<Either<L, TR>> Function(R right) fnR) {
+      FutureOr<Either<L, TR>> Function(R right) fnR) {
     return Future.value(Left<L, TR>(value));
   }
 
@@ -112,8 +114,8 @@ class Left<L, R> extends Either<L, R> {
 
   @override
   Future<Either<TL, R>> thenLeftAsync<TL>(
-      Future<Either<TL, R>> Function(L left) fnL) {
-    return fnL(value);
+      FutureOr<Either<TL, R>> Function(L left) fnL) {
+    return Future.value(fnL(value));
   }
 
   @override
@@ -127,13 +129,13 @@ class Left<L, R> extends Either<L, R> {
   }
 
   @override
-  Future<Either<L, TR>> mapAsync<TR>(Future<TR> Function(R right) fnR) {
+  Future<Either<L, TR>> mapAsync<TR>(FutureOr<TR> Function(R right) fnR) {
     return Future.value(Left<L, TR>(value));
   }
 
   @override
-  Future<Either<TL, R>> mapLeftAsync<TL>(Future<TL> Function(L left) fnL) {
-    return fnL(value).then((value) => Left<TL, R>(value));
+  Future<Either<TL, R>> mapLeftAsync<TL>(FutureOr<TL> Function(L left) fnL) {
+    return Future.value(fnL(value)).then((value) => Left<TL, R>(value));
   }
 
   @override
@@ -161,8 +163,8 @@ class Right<L, R> extends Either<L, R> {
 
   @override
   Future<Either<L, TR>> thenAsync<TR>(
-      Future<Either<L, TR>> Function(R right) fnR) {
-    return fnR(value);
+      FutureOr<Either<L, TR>> Function(R right) fnR) {
+    return Future.value(fnR(value));
   }
 
   @override
@@ -172,7 +174,7 @@ class Right<L, R> extends Either<L, R> {
 
   @override
   Future<Either<TL, R>> thenLeftAsync<TL>(
-      Future<Either<TL, R>> Function(L left) fnL) {
+      FutureOr<Either<TL, R>> Function(L left) fnL) {
     return Future.value(Right<TL, R>(value));
   }
 
@@ -187,12 +189,12 @@ class Right<L, R> extends Either<L, R> {
   }
 
   @override
-  Future<Either<L, TR>> mapAsync<TR>(Future<TR> Function(R right) fnR) {
-    return fnR(value).then((value) => Right<L, TR>(value));
+  Future<Either<L, TR>> mapAsync<TR>(FutureOr<TR> Function(R right) fnR) {
+    return Future.value(fnR(value)).then((value) => Right<L, TR>(value));
   }
 
   @override
-  Future<Either<TL, R>> mapLeftAsync<TL>(Future<TL> Function(L left) fnL) {
+  Future<Either<TL, R>> mapLeftAsync<TL>(FutureOr<TL> Function(L left) fnL) {
     return Future.value(Right<TL, R>(value));
   }